Move ft job list
